separating styles from main file
