+++ App.tsx
@@ -1,20 +1,7 @@
-import { StatusBar } from "expo-status-bar";
-import { StyleSheet, Text, View } from "react-native";
+import Login from './src/screens/login/Login'
 
 export default function App() {
-  return (
-    <View style={styles.container}>
-      <Text>Hello guys welcome</Text>
-      <StatusBar style="auto" />
-    </View>
-  );
+  return <Login />;
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-});
+
