--- src/screens/Login.tsx
@@ -1,91 +1,0 @@
-import React from "react";
-import {
-  StyleSheet,
-  Text,
-  View,
-  SafeAreaView,
-  ScrollView,
-  TextInput,
-  Pressable,
-} from "react-native";
-
-export default function App() {
-  return (
-    <SafeAreaView style={styles.container}>
-      <ScrollView>
-        <Text style={styles.title}>Welcome back</Text>
-        <Text style={styles.subtitle}>We are so excited to see you again</Text>
-        <Text style={styles.text}>ACCOUNT INFORMATION</Text>
-        <TextInput
-          style={styles.input}
-          placeholderTextColor="grey"
-          placeholder="Username"
-        />
-        <TextInput
-          style={styles.input}
-          placeholderTextColor="grey"
-          placeholder="Full name"
-        />
-        <TextInput
-          style={styles.input}
-          placeholderTextColor="grey"
-          placeholder="Password"
-        />
-
-        <Text style={styles.forgotPasswordText}>Forgot password?</Text>
-        <Pressable style={styles.button}>
-          <Text style={styles.buttonText}>Login</Text>
-        </Pressable>
-      </ScrollView>
-    </SafeAreaView>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    backgroundColor: "#36393E",
-    flex: 1,
-    padding: 10,
-    paddingVertical: 30,
-  },
-  title: {
-    color: "white",
-    fontSize: 30,
-    fontWeight: "bold",
-    alignSelf: "center",
-    marginVertical: 10,
-  },
-  subtitle: {
-    color: "lightgrey",
-    fontSize: 20,
-    alignSelf: "center",
-    marginBottom: 30,
-  },
-  text: {
-    color: "white",
-    fontWeight: "bold",
-    marginVertical: 5,
-  },
-  input: {
-    backgroundColor: "#202225",
-    marginVertical: 5,
-    padding: 15,
-    color: "white",
-    borderRadius: 5,
-  },
-  forgotPasswordText: {
-    color: "#4CABEB",
-    marginVertical: 5,
-  },
-  button: {
-    backgroundColor: "#5964E8",
-    alignItems: "center",
-    padding: 15,
-    borderRadius: 5,
-    marginVertical: 10,
-  },
-  buttonText: {
-    color: "white",
-    fontWeight: "bold",
-  },
-});
